requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-vision
google-auth
openai
Pillow
google-cloud-storage
requests
